chore(rewards): move ForcedDarkThemeProvider out of Vip/
It's not VIP-specific — any Rewards surface that wants a forced
dark theme can reuse it. Promote it next to ThemeImageComponent
at the top of Rewards/components/, with its own folder.
