+++ app/components/UI/Rewards/components/ForcedDarkThemeProvider/ForcedDarkThemeProvider.test.tsx
