+++ app/components/UI/Rewards/components/ForcedDarkThemeProvider/ForcedDarkThemeProvider.tsx
